feat: Add config object to DbContext; rename AppContext to ApplicationContext to avoid naming conflicts with existing AppContext class
