+++ Lit.Server/Lit.Server.Api/Controllers/WeatherForecastController.cs
@@ -1,11 +1,14 @@
 using Microsoft.AspNetCore.Mvc;
-
+using Lit.Server.Data;
 namespace Lit.Server.Api.Controllers
 {
     [ApiController]
     [Route("[controller]")]
     public class WeatherForecastController : ControllerBase
     {
+        private readonly ApplicationContext _context;
+        private readonly IConfiguration _configuration;
+
         private static readonly string[] Summaries = new[]
         {
             "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
@@ -13,14 +16,22 @@ namespace Lit.Server.Api.Controllers
 
         private readonly ILogger<WeatherForecastController> _logger;
 
-        public WeatherForecastController(ILogger<WeatherForecastController> logger)
+        public WeatherForecastController(ILogger<WeatherForecastController> logger, ApplicationContext context, IConfiguration configuration)
         {
             _logger = logger;
+            _context = context;
+            _configuration = configuration;
+
         }
 
         [HttpGet(Name = "GetWeatherForecast")]
         public IEnumerable<WeatherForecast> Get()
         {
+            string connectionString = _configuration.GetConnectionString("DefaultConnection");
+
+            // Outputting the connection string to the console
+            Console.WriteLine($"Connection string: {connectionString}");
+
             return Enumerable.Range(1, 5).Select(index => new WeatherForecast
             {
                 Date = DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
