+++ Lit.Server/Lit.Server.Api/Lit.Server.Api.csproj
@@ -14,7 +14,10 @@
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.16" />
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="7.0.0" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
   </ItemGroup>
-
+  <ItemGroup>
+    <ProjectReference Include="..\Lit.Server.Data\Lit.Server.Data.csproj" />
+  </ItemGroup>
 </Project>
